update: require recaptchaToken when login
